style: clear remaining rust-1.97 clippy lints in the grammar-scheduler module

+++ crates/synapse-module/owned-decode-grammar-scheduler/scheduler.rs
@@ -808,7 +808,7 @@ mod tests {
                     saw_release = true;
                 }
                 if arbitration.selected == QueueClass::Decode {
-                    scheduler.commit_quantum("d1", 16, tick as u64);
+                    scheduler.commit_quantum("d1", 16, tick);
                     scheduler.requeue_continuation("d1");
                 }
             }
